fix: deploiement3

+++ src/gable-components/entities/gable-component.entity.ts
@@ -27,7 +27,7 @@ export class Composant extends BaseEntity {
 
 
     @ApiProperty()
-    @Column({ type: 'money' })
+    @Column({ type: 'float' })
     price: number
